Time: 33 ms (64.39%), Space: 94.3 MB (13.77%) - LeetHub

+++ 1438-longest-continuous-subarray-with-absolute-diff-less-than-or-equal-to-limit/1438-longest-continuous-subarray-with-absolute-diff-less-than-or-equal-to-limit.java
@@ -1,23 +1,23 @@
-import java.util.*;
 class Solution {
     public int longestSubarray(int[] nums, int limit) {
+
         Deque<Integer> maxDeque = new ArrayDeque<>();
         Deque<Integer> minDeque = new ArrayDeque<>();
         int left = 0;
         int answer = 0;
         for (int right = 0; right < nums.length; right++) {
-            while (!maxDeque.isEmpty() &&
-                   nums[maxDeque.peekLast()] < nums[right]) {
+            while (!maxDeque.isEmpty()
+                    && nums[maxDeque.peekLast()] < nums[right]) {
                 maxDeque.pollLast();
             }
             maxDeque.offerLast(right);
-            while (!minDeque.isEmpty() &&
-                   nums[minDeque.peekLast()] > nums[right]) {
+            while (!minDeque.isEmpty()
+                    && nums[minDeque.peekLast()] > nums[right]) {
                 minDeque.pollLast();
             }
             minDeque.offerLast(right);
-            while (nums[maxDeque.peekFirst()] -
-                   nums[minDeque.peekFirst()] > limit) {
+            while ((long) nums[maxDeque.peekFirst()]
+                    - nums[minDeque.peekFirst()] > limit) {
                 if (maxDeque.peekFirst() == left) {
                     maxDeque.pollFirst();
                 }
